create, read

+++ src/main/java/com/backend/gini/domain/boards/CategoryRepository.java
@@ -1,4 +1,8 @@
 package com.backend.gini.domain.boards;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+
+import java.util.Optional;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
 }
